HR_Website/requirements.txt
flask
requests
boto3
python-dotenv
pyjwt
